mitre_detect/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
Flask-SSE==0.2.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kafka-python==2.0.1
ksql==0.5.1.1
MarkupSafe==1.1.1
pymongo==3.10.1
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
sse==1.2
urllib3==1.25.8
Werkzeug==1.0.0
xmltodict==0.12.0
